Rework the Gradle build
This commit reworks how the Gradle plugin is built. Instead of using
different versions of Gradle for the purpose of testing (which doesn't
really work because the plugin doesn't do functional testing but uses
the API jar), the same version of Gradle is used (the latest, 6.0.1).

However, this requires a bit of tweaking:

- Gradle will always be executed with a Java 8 JVM, which requires
to set the `JDK_8` environement variable
- when Maven calls Gradle, it tells what is the target bytecode
version, which is then used by Gradle to determine what JDK to use.
It will, for example, use JDK 7 to compile and execute the tests,
using forking. Again, an environment variable `JDK_7` needs to be
set.

The difficulty of this procedure is that this build is building a
Gradle plugin. As such, it relies on the `gradleApi` jar, which is
a generated jar at runtime. This jar contains more than the API,
as the name doesn't imply, because it also contains the minimal
runtime for testing with TestKit.

However, this plugin does _not_ use TestKit. It just runs tests
and checks against the Gradle API. Ideally, this Gradle API should
be published and reusable, and this plugin should be able to be
built for older Gradle versions. This is what the build tries to
emulate by tweaking the Gradle API jar in order to lower its
bytecode version.

This, however, raises different questions:

- if the plugin is built with Java 8, it is likely to depend on
internal details of the API, making it incompatible with previous
Gradle releases, so what's the point of testing?
- if the plugin is built with a previous format, like it does in
this build, then we can execute it on older Gradle releases (likely),
but then we don't test that it will _also_ work with newer JDKs.

For this, Gradle would make it much easier by allowing you to create
different test suites for the different target JDKs. The fact that the
build is driven by Maven doesn't make it easy, but, let's be honest,
the fact that Gradle doesn't publish API jars is also painful.

It would be worth investigating _functional_ testing instead of unit
testing to test the behavior against different Gradle versions, and
in particular look at the https://gitlab.com/ysb33rOrg/gradleTest
plugin.