DRAFT: Optimize transformation on IDE triggered incremental builds.
The Mojo now uses the BuildContext as advertised in [0] to optimize invocations made from m2e which passes the list of changed resources via the BuildContext interface. The draft wraps the Source for the classes to be processed by customizing the iterator to only include classes that are the compiled equivalents of the list of Java files handed into the Mojo via the BuildContext. Also, we signal each of the class files changed back to the IDE via BuildContext.refresh(…).

[0] https://www.eclipse.org/m2e/documentation/m2e-making-maven-plugins-compat.html